refactor based on feedback
